requirements.txt
alembic==1.4.3
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cryptography==3.3.1
ecdsa==0.14.1
fastapi==0.62.0
h11==0.11.0
idna==2.10
loguru==0.5.3
Mako==1.1.3
MarkupSafe==1.1.1
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.20
pydantic==1.7.3
python-dateutil==2.8.1
python-editor==1.0.4
python-jose==3.2.0
python-multipart==0.0.5
SQLAlchemy==1.3.22
starlette==0.13.6
urllib3==1.26.2
uvicorn==0.13.1
